+++ ChangeLog.md
@@ -272,3 +272,4 @@
   * Fix syntax error
 - 0.48.7
   * Fix an Error that is thrown while parsing a RSS file: Can't parse items without a publication date
+	* Sort Source now case insensitive